requirements.txt
python-dotenv
pandas
numpy
openai
chardet
Flask
Werkzeug
